+++ src/main/scala/com/wavesplatform/Application.scala
@@ -1,6 +1,7 @@
 package com.wavesplatform
 
 import akka.actor.Props
+
 import com.wavesplatform.consensus.WavesConsensusModule
 import com.wavesplatform.http.NodeApiRoute
 import scorex.account.AddressScheme
@@ -10,12 +11,12 @@ import scorex.consensus.nxt.api.http.NxtConsensusApiRoute
 import scorex.network.{TransactionalMessagesRepo, UnconfirmedPoolSynchronizer}
 import scorex.utils.ScorexLogging
 import scorex.waves.http.{DebugApiRoute, ScorexApiRoute, WavesApiRoute}
-import scorex.waves.settings._
+import com.wavesplatform.settings._
 import scorex.waves.transaction.WavesTransactionModule
 
 import scala.reflect.runtime.universe._
 
-class Application(val chainParams: ChainParameters, appSettings: WavesSettings) extends {
+class Application(appSettings: WavesSettings) extends {
   override implicit val settings = appSettings
   override val applicationName = "waves"
   override val appVersion = {
@@ -27,7 +28,7 @@ class Application(val chainParams: ChainParameters, appSettings: WavesSettings) 
 
   override implicit lazy val consensusModule = new WavesConsensusModule()
 
-  override implicit lazy val transactionModule = new WavesTransactionModule()(settings, this, chainParams)
+  override implicit lazy val transactionModule = new WavesTransactionModule()(settings, this, settings.chainParams)
 
   override lazy val blockStorage = transactionModule.blockStorage
 
@@ -73,20 +74,44 @@ class Application(val chainParams: ChainParameters, appSettings: WavesSettings) 
 }
 
 object Application extends App with ScorexLogging {
-  log.debug("Start server with args: {} ", args)
+  //TODO: gagarin55-change to info cuz default log level is info
+  log.debug("Starting with args: {} ", args)
 
   private val filename = args.headOption.getOrElse("settings.json")
   private val settings = new WavesSettings(filename)
-  private val chainParams: ChainParameters = if (settings.isTestNet) TestNetParams else MainNetParams
+
+  configureLogging(settings)
 
   // Initialize global var with actual address scheme
-  AddressScheme.current = chainParams.addressScheme
+  AddressScheme.current = settings.chainParams.addressScheme
 
-  val application = new Application(chainParams, settings)
+  log.info(s"${Constants.AgentName} Blockchain Id: ${settings.chainParams.addressScheme.chainId}")
+
+  val application = new Application(settings)
   application.run()
-
-  log.debug("Waves has been started")
 
   if (application.wallet.privateKeyAccounts().isEmpty)
     application.wallet.generateNewAccounts(1)
+
+  /**
+    * Configure logback logging level according to settings
+    */
+  def configureLogging(settings: WavesSettings) = {
+    import ch.qos.logback.classic.LoggerContext
+    import org.slf4j._
+    import ch.qos.logback.classic.Level
+
+    val lc = LoggerFactory.getILoggerFactory.asInstanceOf[LoggerContext]
+    val rootLogger = lc.getLogger(Logger.ROOT_LOGGER_NAME)
+    settings.loggingLevel match {
+      case "info" => rootLogger.setLevel(Level.INFO)
+      case "debug" => rootLogger.setLevel(Level.DEBUG)
+      case "error" => rootLogger.setLevel(Level.ERROR)
+      case "warn" => rootLogger.setLevel(Level.WARN)
+      case "trace" => rootLogger.setLevel(Level.TRACE)
+      case _ =>
+        log.warn(s"Unknown loggingLevel = ${settings.loggingLevel}. Going to set INFO level")
+        rootLogger.setLevel(Level.INFO)
+    }
+  }
 }